requirements.txt
aiosqlite==0.18.0
alembic==1.10.2
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.93.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
orjson==3.8.7
passlib==1.7.4
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.6
pyflakes==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.25.0
text-unidecode==1.3
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
